# Improve StringBuilder ctor(), ctor(int), and ToString() performance.