Squeeze a little more performance out of StringBuilder ctor, ToString + others.

- The default `ctor()`, and by far the most used, goes through 9 conditions, a string pin, and a function call. None of which are needed.
Changed implementation to be just the two things that are required.
- The `ctor(int)`, second most common, also does a string pin and function call. Redirect to the smaller `ctor(int, int)` overload.
- Missing `(uint)` cast caused additional casting code to be generated.
- Fix on `&array[0]` instead of `array` for less generated IL. Lengths are checked beforehand.
- Moved return statements inside the `fixed` block so that the `char*` doesn't need to be nulled out.

| Test | Before | After |
------ | ------- | --------
| ctor() | 173ms | 93ms |
| ctor(16) | 149ms | 109ms |
| ToString() | 110ms | 93ms |

Perf test code https://gist.github.com/bbowyersmyth/2cc93ec0b2127cd98d567654b5d6467e
Additional test https://github.com/dotnet/corefx/pull/11353